Major cleanup of the AsPath and related classes: Use a single class for AS_SET and AS_SEQUENCES, simplify the internal representation, fix printing, remove some redundant methods, etc. etc.
The code passes all checks though there are surely some memory leaks
as a result of a change in AsPath::encode() routine that I am going
to address next when revising the Packet class.

This code might still enjoy some work -- e.g. the external
representation of AsPaths (the one used in packets) is quite compact
and fast and easy to parse, which makes me wonder why we are not
using the same one (or something very similar) for internal
representation.

Also, the way the code is structured now, it makes a lot of
possibly unnecessary memory allocations and copies. Once the Packet
class has been revised, we might try to optimize things so that we
construct packets in preallocated buffers.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@28 26435e26-4b25-483a-942d-477236992e84