Re-think how we store IGP metrics for MED insertion.
We now store the IGP metric that is used by Decision in the subnet
route that wins.  This uses more storage than the previous version,
but it means the metric that was used will still be around when the
route is later replaced.  This will eventually allow RibOut to
suppress sending changes to peers if the final route we were about to
send hasn't actually changed (because for example, the route changed
internally due to the MED changing, but we removed the MED from this
peering, so externally the route is unchanged).


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@46 26435e26-4b25-483a-942d-477236992e84