Rework the destructor for BGPMain. Previously calling the rib_ipc_handler destructor would cause the ebgp and ibgp tables to be deleted from the RIB. Unfortunately the callback would be occur on the freed rib_ipc_handler.
Now the table deletion is done explictly from the BGPMain destructor. The
rib_ipc_handler no longer attempts the table deletion in the destructor but
will generate a warning if the table exist.

Fixed a bug in the rib_ipc_handler that stopped the deletion working,
the RIB name was being cleared and then used.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@85 26435e26-4b25-483a-942d-477236992e84