Fix the mechanism for deletion of PimMre entries: If an entry is marked for deletion, then schedule a task that will adjust the pointers from other related PimMre entries to that entry. When the task is executed, it will check first whether the entry is still a candidate for removal (e.g., if there was a Join message received right before this task was run). Only if the entry is still a candidate for removal, then pull it out of the multicast routing lookup table, and adjust the pointers to that entry. After all entries scheduled for removal with the same task have been processed and pulled out, then free the memory they are using.
Previously, a PimMre routing entry was pulled-out of the lookup
table right when the "PimMre delete" task was scheduled, and then
the task had only to adjust the pointers to it and to free its
memory.
However, this created the race condition that there might be
already some other tasks (scheduled to run before the
"PimMre delete" task) that may be refering to (and modifying)
this entry. If the PimMre entry was already pulled-out of
the multicast routing lookup table, then those tasks will be
modifying "half-dead" PimMre entry, and there will be
some loss of information.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@208 26435e26-4b25-483a-942d-477236992e84