Bugfix in assignment to a local static variable in method PimVif::upstream_join_timer_t_override()
Originally the code was:
  static struct timeval timeval = vif_override_interval();

However, 'timeval' was suppose to be assigned every time the method
is called, which will not happen because 'timeval' is static (i.e.,
it will be assigned only the first time the method is called).
The solution is to explicitly perform the assigment on a separate
line.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@301 26435e26-4b25-483a-942d-477236992e84