Add cref_ptr - compact reference pointer class.  A slower, but more compact reference pointer class with exactly the same interface as the existing ref_ptr.
A cref_ptr is 4 bytes big.  It represents an index into a table.  Each
entry in the table contains a 4 byte counter and a void* pointer.  The
entry is shared by however many cref_ptr's reference the underlying
data.

The penalty of cref_ptr is a function call and an indirection when
accessing the data.  In contrast, ref_ptr has the pointer as part of
the structure and has no access penalty, but is larger at 4 bytes plus
the size of the pointer.

On a 32-bit box with n cref_ptr's to same data the size is 8 + 4n
bytes, whilst for a ref_ptr it's 4 + 8n bytes.  We could squeeze the
cref_ptr size of cref_ptr down further, but the current implementation
will cause limit the number of available cref_ptr's accordingly (see
code to understand why).


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@376 26435e26-4b25-483a-942d-477236992e84