Change the mechanism to recompute CouldAssert(S,G,I), CouldAssert(*,G,I), AssertTrackingDesired(S,G,I), and AssertTrackingDesired(*,G,I):
Originally, PimMre::recompute_assert_tracking_desired_sg(vif_index)
and friends were applied per interface, and that interface
was the one that had triggered the recomputation (e.g., because
a Join/Prune, etc message was received on that interface).
However, in case of events such as receiving winning Assert message,
Join/Prune, etc on an interface, this may change the
AssertTrackingDesired(S,G,I) state of the RPF_interface(S) or
RPF_interface(RP(G)); i.e., the state of an interface that is
different from the one that had triggered the recomputation.

The solution is whenever there is a single event that may affect
the AssertTrackingDesired(S,G,I) state for any interface, then
recompute AssertTrackingDesired(S,G,I) for all interfaces in
parallel, and apply the appropriate actions only for those
interfaces whose state has changed.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@403 26435e26-4b25-483a-942d-477236992e84