A coredump fix:
Add one more argument to NextHopRibRequest<A>::register_interest_response()
callback: nexthop_interest (the address we are attempting to register
with the RIB). This address is needed in case we need to attempt again
to register same address in case of failure. Otherwise, using (*addr)
for that (the original solution) may coredump because "addr" will
be NULL in case of RESOLVE_FAILED for example.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@1056 26435e26-4b25-483a-942d-477236992e84