If an XRL is received for unexpected address family, then silently ignore it instead of printing a message.
The reason is because when the MFEA that may be either IPv4 or
IPv6-only capable registers with the FEA, the registration is not
IPv4 or IPv6-specific, therefore the FEA will send interface address
information about IPv4 AND IPv6 addresses. Hence, the MFEA
should silently ignore the address family info it doesn't expect.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@1121 26435e26-4b25-483a-942d-477236992e84