Add shutdown XRL to common interface.  In many processes this is implemented merely as a placeholder.
The rtrmgr's expectation is that if you return OKAY in response to
shutdown, that you'll then be in SHUTDOWN state when it queries with
get_status.  If not, it'll assume something failed, and kill you.  But
it won't try shutdown unless the XRL template file says to do so.  At
least that's the theory - the rtrmgr part isn't implemented yet.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@1262 26435e26-4b25-483a-942d-477236992e84