Fix a Linux netlink-related problem.
When we receive a netlink message from Linux netlink socket, we
should probably check that nlmsg_pid of the received message
matches our pid.
However, it appears that on return Linux doesn't fill-in
nlh->nlmsg_pid to our pid (e.g., it may be set to 0xffffefff).
Unfortunately, Linux's netlink(7) is not helpful on the
subject, hence we ignore this additional test.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@1338 26435e26-4b25-483a-942d-477236992e84