Fix a Netlink-related problem for Linux.
Unlike routing sockets which when opened with AF_UNSPEC as the
third argument have dual IPv4/IPv6 status, we cannot open a netlink
socket that is dual (i.e., the third argument must be either
NETLINK_ROUTE or NETLINK_ROUTE6).

Hence, added two new classes NetlinkSocket4 and NetlinkSocket6
that use the existing NetlinkSocket class as a base class.
Then, a NetlinkObserver constructor is created with two arguments:
ns4 and ns6, so it is observing both IPv4 and IPv6 netlink sockets.
When data is received on either ns4 or ns6, it is passed to the
same observer, and then processed as necessary. The data processing
was already dual IPv4/IPv6, hence we don't really care to know
whether the data came from ns4 or ns6.

The IPv4 code appears to work; the IPv6 code is untested
(though, I am not sure whether NETLINK_ROUTE6 is really
implemented in the kernel).

After this fix, pim/test_pim appears to work on Linux as well.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@1343 26435e26-4b25-483a-942d-477236992e84