The handling of XRL errors should now correspond to the error document, including checking the error returns from the XRL sends.
Retransmit XRLs when a timeout occurs.

The code to support more than one XRL in flight did not work fixed
code and we now allow 100 XRLs in flight.

The sendit method was specialised for IPv4 and IPv6, unfortunately
some changes had only made it to the IPv4 method. Unspecialised the
sendit method and added two smaller specialised methods sendit_spec.

The MAX_ERR_RETRIES was previously set to 30, take its value from the
FLYING_LIMIT. When there is a glitch, this allows all XRLs in flight to
return REPLY_TIMED_OUT.

Now that we support retransmissions on REPLY_TIMED_OUT errors it is
possible to send the same add/delete request to the RIB. Currently a
duplicate request to the RIB returns COMMAND_FAILED, which BGP
considers a fatal error. Perhaps the error return should supply the
reason for the COMMAND_FAILED. Conversely the point may be moot, when
we move to reliable XRLs the REPLY_TIMED_OUT error will no longer
occur.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@1508 26435e26-4b25-483a-942d-477236992e84