Added two new XRLs to the FEA interface manager interface: get_all_vif_pif_index get_configured_vif_pif_index
Those XRLs are needed by the MFEA to get the physical interface
index of an interface. The MFEA itself requires the physical interface
index (as known by the kernel) in case of IPv6 to setup multicast virtual
interfaces, etc.
Other modules (RIB, protocols, etc) that do not operate directly on the
kernel and/or may be running on a host that is different from the
forwarding plane do not need this pif_index.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@1549 26435e26-4b25-483a-942d-477236992e84