Redesign the handling of the IGMP/ICMPV6 socket and the socket used for multicast routing.
In case of FreeBSD, we can open two IGMP sockets: if we set one
as a socket for multicast routing, then we can use both sockets
to receive IGMP data messages.
However, in case of Linux, we would receive the IGMP Membership
Report messages only on the mrouter socket. In other words, the
second IGMP socket will NOT receive the IGMP Membership Report
messages.

The fix (applied by this commit) to the above problem is to use
a single socket that is shared for both IGMP/MLD protocol messages
AND for multicast routing.
The sharing mechanism takes care of properly opening the socket
if it wasn't open by any party, and to close it when
is not used by both parties anymore.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@1826 26435e26-4b25-483a-942d-477236992e84