Add new methods to FtiConfig to get the status of the kernel whether it was enabled to perform unicast forwarding, and methods to enable/disable unicast forwarding:
- unicast_forwarding_enabled[4,6]
- accept_rtadv_enabled6
- set_unicast_forwarding_enabled[4,6]
- set_accept_rtadv_enabled6

Currently, the inner code of those methods uses sysctlbyname() and
is #ifdef for all supported OS-es: *BSD and Linux.
Later, when we have support for more OSes, we must add the appropriate
method for each new platform, otherwise the above code will
fail to compile with #error

The second and the fourth methods are needed for *BSD
to get/set the flag regarding acceptance of IPv6 Router Advertisement
messages, and are no-op for Linux.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@1935 26435e26-4b25-483a-942d-477236992e84