Reverse the functions of 'j' and 'k' in page mode:  - 'j' scroll down one line  - 'k' scroll up one line
This behavior is now consistent with "vi" and "more/less" key
binding.
If in the future we want to be consistent with the CLI behavior of
some router vendors which use different key binding from the key
binding of "vi/more/less", then we can add the hook later.

Requested by Atanu.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@1997 26435e26-4b25-483a-942d-477236992e84