Add "replay" support.
All update packet sent to or by a test_peer are saved in a
corresponding trie. Each trie is capable of generating the minimum list of
update packet that created the state in that trie (including update
packets containing only withdraws). This replay state can now be
extracted from a trie.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@2185 26435e26-4b25-483a-942d-477236992e84