Change list<string> argument in XrlAction::check_xrl_is_valid() to a const reference to avoid a gratuituous copy.
Use a string rather than a dynamically character array for cleaned_xrl
in XrlAction::check_xrl_is_valid(). It's generally safer and makes
more sense in the context (ie the methods that are supplied with the
cleaned_xrl take a const string reference so this avoids creating a
temporary).

s/u_int/uint32_t/


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@2378 26435e26-4b25-483a-942d-477236992e84