Change IPNet<A>::_prefix_len type from size_t to uint32_t.
Propagate change such that uint32_t is used for all IP bit related
operations.  This avoids any size_t <-> uint32_t conversions.  NB
arguments and return types for operations relating to to byte sizes
remain of type size_t.

Rationale: size_t can be 64bits on a 64bit platform, which is
excessive for an 8bit number.  Whilst size_t is defined to be unsigned
in C99, there are platforms that still use a signed value.  This makes
format strings difficult since %z format argument is not widely enough
deployed to be useful today.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@2388 26435e26-4b25-483a-942d-477236992e84