If we have netlink sockets, then we unconditionally set HAVE_NETLINK_SOCKETS_SET_MTU_IS_BROKEN and HAVE_NETLINK_SOCKETS_SET_FLAGS_IS_BROKEN to 1.
The reason is because if we attempt to set the MTU or the interface
flags on a Linux system (e.g., Linux RedHat-7.x with kernel 2.4-x),
this is a no-op: nothing happens, but the kernel doesn't return
an error. The test whether we can really set the MTU and the interface
flags in the kernel is rather complicated, and would require root
privileges.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@2418 26435e26-4b25-483a-942d-477236992e84