We have four separate instances of routing tables in BGP: IPv4 (AFI 1) Unicast (SAFI 1) IPv4 (AFI 1) Multicast (SAFI 2) IPv6 (AFI 2) Unicast (SAFI 1) IPv6 (AFI 2) Multicast (SAFI 2)
When an update packet is teased apart and each route is injected into
the plumbing it is obvious where it should be injected. When the
message pops out of the plumbing at the peer handler the AFI
information is preserved as its implemented using templates. The SAFI
information, however is lost.

The simplest way of preserving the SAFI information is to place it in
the subnet route message. I tried this but in some methods
it became awkward.

Therefore the relevant peerhandler methods now take a SAFI argument
and each instance of the ribout is passed a SAFI in its constructor.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@2544 26435e26-4b25-483a-942d-477236992e84