1) The BGP process registers interest in a nexthop. 2) The route which resolves this nexthop is deleted, causing an invalid route being sent to BGP.
The BGP process should see the register response followed by the
invalidate.  Unfortunately sometimes the invalid may arrive at the BGP
process before the register interest response. This behaviour is an
artifact of the XRL implementation.

In the case were we see the invalidate before the register response:
We save the invalidate and when the register response arrives we
verify they match then rerequest the register response.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@3015 26435e26-4b25-483a-942d-477236992e84