BGP processes update messages by passing them through a series of route tables. The route tables are mostly statically plumbed together. There are however two dynamic tables the "dump table" and the "deletion table". When a session is established the "dump table" is responsible for passing all currently winning routes to this new peer. Once all the routes have been sent to the peer the "dump table" removes itself from the plumbing. The "deletion table" is used to remove routes when a session goes down. The "deletion table" also removes itself when all the routes sent by this peer have been removed.
There are a number of issues with the interactions between the two
tables. The specific problem we are trying to address is the case of a
"deletion table" existing and hence sending deletes when a new peering
is established. Previously when a new peering was established the
"dump table" would be made aware of all live sessions, but not of any
of the sessions which were down but still sending deletes. If a
session goes down (causing a "deletion table" to be installed) while
another another session is being brought up (by a "dump table"). Then
the "dump table" has to deal specially with the deletes being sent by
the "delete table". The "dump table" has to either supress or pass
through the deletes, but most importantly it has to stay around until
the "deletion table" has finished sending deletes. It is important to
note that any unneccesary deletes that make it to the downstream cache
will cause an XLOG_ASSERT. As deletes must have corresponding state
in the cache.

Previously a new session being brought up while a deletion was in
progress would fail. As the new peer was unaware of the downed peer
any deletes that came from this peer would appear to be from a new
peer. So the "dump table" would pass them through and the cache would
signal an inconsistency by calling XLOG_ASSERT.

The fix is to tell the "dump table" about any "delete tables" that are
running when the "dump table" is instantiated. This fixes all but one
of the regression tests.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@3048 26435e26-4b25-483a-942d-477236992e84