Add a first-cut (incomplete) implementation of the Fib2mrib module. Currently, the code is almost a copycat of the StaticRoutes module.
The purpose of the Fib2mrib module is to get the unicast forwarding
routes from the kernel (via the FEA) and add them to the MRIB so they
can be used for multicast RPF purpose. We need this in case we
are not running any uncast routing protocols and/or if the router
is running non-XORP unicast routing protocols.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@3086 26435e26-4b25-483a-942d-477236992e84