Back-up an earlier changes that removes the PimMfe entry if there are no outgoing interfaces.
Removing immediately the PimMfe entry creates a problem in
case of DR with directly-connected source that has been
surpressed by PIM Register-Stop messages.
In that case, we want to have the corresponding MFC entry
in the kernel to block all the multicast traffic.
Otherwise, the PimMfe entry is removed, then immediately installed,
then immediately removed, and so on.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@3188 26435e26-4b25-483a-942d-477236992e84