Bug fix:
Handle properly inside Mrt::source_by_prefix_end()
and Mrt::group_by_prefix_end() the corner case
when the prefix address is same as a network mask.
E.g.: 255.0.0.0/8 .
The problem is rooted in the fact that
INCR(255.0.0.0/8) is (0.0.0.0/8).

In that case, the correct result to return is the end of the table
rather than the lower-bound for 0.0.0.0/8 (e.g., somewhere in
the beginning of the table).


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@3230 26435e26-4b25-483a-942d-477236992e84