When a new peering comes up we iterate through the RIB-INs pulling routes through the system. While we are dealing with existing routes a new route may appear. If the new route is behind the iterator then we deal with it. If it is in front we drop it as we know we will reach this route eventually.
Routes were coming through in front of the iterator but the iterator
never seemed to find them. The problem was the iterator was not
pointing at the current entry but the next entry, noticed by mjh.

The iterator know points at the current entry a side effect is that we
have to make one more call to the dump code to discover we have
finished. The test code and reference file have had to be changed to
reflect this change.

The test "test_peering2.sh -t test3" should hopefully stop failing
now.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@3334 26435e26-4b25-483a-942d-477236992e84