Fix an unexpected fatal scenario which may happen if we delete an interface on Linux:
In case of Linux kernel we may receive first a message to delete the
interface and then a message to delete an address on that interface.
However, the first message would remove all state about that interface
(including its addresses).
Hence, we silently ignore messages for deleting addresses if the
interface is not found.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@3455 26435e26-4b25-483a-942d-477236992e84