Added new RIB XRLs:
add_interface_route[4,6]
replace_interface_route[4,6]

Those XRLs are used to add or replace a route by explicity specifying
the network interface/vif toward the destination instead of relying
on the RIB to calculate the interface toward the next-hop router
(based on subnet address information).
For example, in case of wireless protocols we cannot use subnet addresses
to compute the wireless interface because in general the neighbors
may not share a common subnet address.

The existing XRLs add_route[4,6] and replace_route[4,6] are kept unmodified.

Note that for the time being there are no delete_interface_route[4,6]
XRLs simply because we need only the destination prefix address
to identify and delete an existing route.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@3478 26435e26-4b25-483a-942d-477236992e84