Remove XLOG_FATAL() in cae the kernel did not provide the interface name as an IFLA_IFNAME attribute of an RTM_NEWLINK message.
It could be because this is a wireless event carried to user
space. Such events are encapsulated in the IFLA_WIRELESS field of
a RTM_NEWLINK message.
Unfortunately, there is no easy way to verify whether this is
indeed a wireless event or a problem with the message integrity,
hence we silently ignore it.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@3479 26435e26-4b25-483a-942d-477236992e84