When received a packet, if we cannot find the vif the packet was received on, then ignore the packet rather than coredump.
This is to avoid the race condition when a packet is received on
an interface, but then the interface is removed immediately
(e.g., if a tunnel).
However, we still could receive the XRL with the interface removal
instruction before the XRL with the packet itself, therefore
we should be prepared against that.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@3542 26435e26-4b25-483a-942d-477236992e84