rtrmgr config file save hook now works.
If a program is specified on the xorp_rtrmgr command line using the -s
flag, then this program will be run after the xorp config file is
saved.  The first parameter is the filename of the saved config file.
The userid used to execute the program is that of the user that
requested the rtrmgr to save the config, and the GID is that of group
"xorp".

This is intended to be used when running XORP from a LiveCD, where any
files saved to the memory filesystem would otherwise be lost.  The
program run by the save hook can preserve the config files in stable
storage (such as on a floppy), and when the router restarts, the
preserved files can be restored.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@3775 26435e26-4b25-483a-942d-477236992e84