re-write final completion and self-deletion behaviour of dump table.
The jist of the problem is that when the dump is complete, the
DumpTable has to remain around until all DeletionTables that were in
place at any time in the dump have finished, or we can't ensure
downstream consistency.

When the last of those DeletionTables goes away, DumpTable previously
deleted itself.  BUT it can't do so.  There may be deleted routes in
the FanoutTable queue.  So we need to record that it can now delete
itself, and only do so when the queue upstream is finally empty.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@3807 26435e26-4b25-483a-942d-477236992e84