Add support for "%modinfo: startupmethod" to specify the optional XRL that should be sent to a process once it has started to explicitly activate it. Currently, the "startupmethod" XRL is sent once the process is in PROC_STARTUP state.
The reason we sometime need this explicit activation XRL is
when more than one modules are linked within the same binary.
For example, the xorp_fea binary contains the MFEA. However,
currently even if the MFEA is not needed (e.g., multicast is not
configured), then the xorp_fea will still have to call the MFEA
startup() method to activate it so its state becomes PROC_READY
(just in case it is needed) which is wrong.
With the explicit "startupmethod", the MFEA can stay in PROC_STARTUP
state until it receives this explicit startup XRL.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@3832 26435e26-4b25-483a-942d-477236992e84