Xorpsh operation command template related changes:
Implement support for specifying positional arguments
to the binary command. E.g., $0 is replaced with all
arguments, $1 is replaced with the first argument,
$2 is replaced with the second argument, etc.

For example, originally if the template is:

show interfaces $(interfaces.interface.*) {
    %command: "fea/tools/show_interfaces" %help: HELP;

then when "fea/tools/show_interfaces" is called, xorpsh
will call it with three arguments: "show", "interfaces",
and the particular interface name.

With the new positional argument support the template
%command should explicitly contain $0 to suppy all arguments
as above:
show interfaces $(interfaces.interface.*) {
    %command: "fea/tools/show_interfaces $0" %help: HELP;

If we want to call "show_interface" with argument "-i <iface_name>",
then the template should be:

show interfaces $(interfaces.interface.*) {
    %command: "fea/tools/show_interfaces -i $3" %help: HELP;


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@3908 26435e26-4b25-483a-942d-477236992e84