Add new output state to the state-dependency tracking machinery: OUTPUT_STATE_UPDATE_SPTBIT_MFC.
This state is used on PimMre entries to recompute and eventually
set the SPT bit for the corresponding (S,G) entries.

Note that the spec says that the recomputation
by calling Update_SPTbit(S,G,iif) should happen when a packet
arrives. However, given that we do not forward the data packets
in user space, we need to emulate that by triggering the
recomputation whenever some of the input statements have changed.
Also, note that this recomputation is not needed if there is not
underlying PimMfc entry; the lack of such entry will anyway trigger
a system uncall once the first multicast packet is received for
forwarding.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@3945 26435e26-4b25-483a-942d-477236992e84