Remove code in PimMfc::recompute_iif_olist_mfc() that unnecessary removes the PimMfc entry if its old olist was NULL.
Originally, the code was needed so once the PimMfc entry is
deleted, then the NOCACHE upcall will not only recreate it,
but will also trigger recomputation of the SPT bit for the
corresponding (S,G) entry.
However, the side effect of removing a PimMfc entry was that
the (S,G) KeepAlive timer was automatically cancelled.
The KeepAlive timer cancellation would eventually affect
the (S,G) routing state and would result in other various side
effects and undesired behavior.

Now this code can be removed, because the state-dependency
tracking machinery takes care of the SPT bit so we don't
need to re-create the PimMfc entry just to update that bit.

[Bug report by Marian Durkovic <md@bts.sk>]


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@3946 26435e26-4b25-483a-942d-477236992e84