Don't propagate interface-related changes that just appear beneath us within the system. I.e., all changes should be through the FEA.
This fix is needed to address the issue of keeping (to some extent)
the FEA interface-related internal state in-sync with
the rtrmgr configuration. In particular, if an interface address
is marked as disabled within xorpsh, then the FEA will mark
it as disabled within its own copy of the configuration tree,
but the address will be removed from the kernel. The address removal
however will trigger a routing socket/netlink message to pop-up
to user space to the IfConfigObserver. If those messages are
unconditionally propagated by LibFeaClientBridge, then
the state seen by interested parties (the protocols) will be different
from the state within the FEA.

A better solution would be that all IfConfigObserver changes
should be propagated through the FEA internal configuration state,
and the logic for maintaning that state will decide whether to
apply the changes and propagate them further or to throw them away.

Approved by: Atanu


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@4060 26435e26-4b25-483a-942d-477236992e84