Install all the hooks so that when keyboard interrupts are propogated to the xorpsh cli we can terminate any outstanding operational commands.
Get a handle to the operational command so if the user interrupts the
xorpsh the command can be terminated.

Modify popen2() to return the process ID so we can kill the process if
required.

Save the process ID in the OpInstance. Call pclose() so we don't leak
file descriptors anymore.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@4081 26435e26-4b25-483a-942d-477236992e84