The RIB registers interest with the FINDER regarding the status of processes that are feeding it routes.
When a protocol such as BGP dies the RIB will be notified. The RIB
will then remove all the routes that have been supplied by BGP.

Insert a XLOG_INFO so there is an entry in the logfile when the RIB
believes a protocol has died.

Occasionally the RIB will be sent a delete_route command and the route
will not be present:
[http://www.xorp.org/bugzilla/show_bug.cgi?id=62]

1) BGP fails to call "eventloop.run()" frequently enough.
2) The FINDER concludes BGP is dead and notifies the RIB.
3) The RIB starts to delete the the routes it received from BGP.
4) BGP comes back to life and sends a delete for a route it had
previously added.
5) The RIB no longer has this route oops.

If this really is the problem then we should see this entry in the
logfile along with the failed delete.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@4234 26435e26-4b25-483a-942d-477236992e84