Add protocol_origin field to the following XRLs. redist_transaction4/0.1/delete_route redist_transaction6/0.1/delete_route
The protocol_origin field is needed so an entity that filters
add_route operations based on the protocol origin can filter
the delete_route operations as well. This is the case with
the FEA which needs to filter all add/delete operations
with protocol_origin of "connected".

Bugzilla URL:	http://www.xorp.org/bugzilla/show_bug.cgi?id=85


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@4439 26435e26-4b25-483a-942d-477236992e84