Update the mechanism for registering Get/Set/Observer mechanisms for handling the network interfaces, the forwarding entries and the forwarding table. Previously, there could be only a single mechanism per operation.
Now, we can have a list of mechanisms per operation.
This change is needed so we can modify at the same time two or more
semi-independent entities; e.g., the kernel and Click.
For example, to add a network interface, we have to enable/configure
it in the kernel and inside Click.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@4494 26435e26-4b25-483a-942d-477236992e84