Added a mechanism for the rtrmgr to restart a process that has unexpectedly died.
Currently, the mechanism is based on observing the child process
status of a processes that was forked by the rtrmgr.
However, if the child process was killed by SIGTERM or SIGKILL,
then that process is NOT restarted; the assumption is that
the user really wanted that process killed.

However, the restarting does not always work properly. E.g.,
if a process fails during reconfiguration via xorpsh, then
the rtrmgr itself may coredump. One of the reasons for the
coredump is because there are outstanding XRLs that haven't
been completed (or failed) when the child process failed,
and restarting is possible only if the previous state was clean.
Fixing the problem requires some non-trivial changes to the rtrmgr.
Therefore, the restarting mechanism is disabled by default,
and can be enabled by the new "-r" rtrmgr command-line option.

Ideally, we would use the finder to observe the XRL status of those
processes, therefore in the future this mechanism may be removed.

The code for restarting failed processes is based on a patch
provided by "Ray Qiu <ray.qiu@gmail.com>"

Submitted by:	Ray Qiu <ray.qiu@gmail.com>


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@4501 26435e26-4b25-483a-942d-477236992e84