Exclude compilation flag -Wnon-const-format, because it is problematic.
This flag appears to exist only on earlier versions of gcc
(e.g., gcc-2.95.x), and only on FreeBSD, but not on Linux.
In addition, this flag is problematic, because if it is enabled we cannot
use statements like "vfprintf(stderr, fmt, ap);" where "fmt"
is a variable (e.g., defined as "const char *").


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@4505 26435e26-4b25-483a-942d-477236992e84