bgp/harness/test_rib1.sh -t test9 Was occasionally failing on multiprocessor Linux box (koala). This is a simple test, a route is originated by BGP for which the nexthop does resolve. A route is added then immediately deleted from the RIB. This sequence of events will cause the nexthop to resolve and a BGP add route should be generated followed by a BGP withdraw route. The test does not check for the generation of the packet only that BGP is still running at the end of the test. Unfortunately BGP was failing.
	BGP						RIB

					 		<- route_info_invalid4
	deregister_interest4 ->

A route add followed by an immediate route delete is sent to the
RIB. The RIB sends a message to BGP telling it that the route which
has been deleted from the RIB is invalid. At the same time BGP is
de-registering interest in this route. By the time the de-register
reaches the RIB it no longer has a registration, it returns an error
to BGP which considers any error fatal.

The BGP process is unexpectedly de-registering interest in this next
hop because in RibInTable<A>::push_next_changed_nexthop(), a replace
route has been replaced with a delete route followed by an
add_route. The delete route causes this de-registration.

This condition race has always existed.

The nexthop code was already storing any premature invalidates that it
received. When a de-register fails check the premature invalidate
state.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@4516 26435e26-4b25-483a-942d-477236992e84