Modify method IfConfigSet::set_interface_flags() to include an extra argument "is_up" that indicates whether the interface is UP.
Note that this flag is redundant, because the same information
can be obtained by (flags & IFF_UP), but such redundancy eliminates
inter-mechanism knowledge. E.g., if we want to use (flags & IFF_UP)
in case of Click FEA implementation, we make the assumntion that
the Click FEA implementation has knowledge that the flags
always have IFF_* meaning.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@4592 26435e26-4b25-483a-942d-477236992e84