Provide a definition of take_table_ownership() for the IPFW provider subsystem.
Add a comment about moving the method for converting
rules from XORP to IPFW to be closer to the IpfwFwRule
class itself.

Add specialized template functions to fulfil the above
objective, to catch up with changes to the header file.
Define only IPv4 and IPv6 specializations as other
conversions make no sense.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@4620 26435e26-4b25-483a-942d-477236992e84