Experimental support for policy filtering in RIP.
Route pushing will cause the route db to be flushed.

Other than route filtering, the following has been added:
If a route is learnt from the RIB [route redist], it is saved.
If a route from the RouteDB is deleted, the saved RIB routes are checked
to see whether a RIB route alternative is available. If so, it is added to the
RouteDB. In the previous implementation, it seemed that RIB routes would
"get lost" if they were replaced somehow (for example an external advertisement
with a lower cost was received).

Two of the RIP tests may fail on slow boxes due to filtering and this new
RIB routes check. These tests rely on completion occuring in at most X amount of time.
Due to the new code, things take longer.
The solution may be
1) Changing X
2) Making my code faster
3) Have a re-think about the tests

The tests involved are: test_route_walk and test_timers


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@4628 26435e26-4b25-483a-942d-477236992e84