There are two issues:
1) users may not enter + or special characters in the configuration file
as they were not properly escaped when the XRL is constructed.

2) The rtmgr does not escape nested quotes when sending its configuration
to the xorpsh.

Both of these were fixed in a quick/temporary way (not ideal).

The policyvarmap.conf is a configuration which the policy process must
read at boot up. A mechanism via which this may be done has to be thought of.
Currently, the policy process will simply fopen it.

Protocols wishing to support policy filtering will need to edit the conf file.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@4633 26435e26-4b25-483a-942d-477236992e84