Added configuration directives for policies.
Protocol use import "policy1,policy2,policy3" to filter the input branch.
export "policy1,policy2,policy3" to filter the output branch and to enable
route redistribution.

The old redistribution interface for rip is hidden by the new one, although
still present in the code.

An initial configuration template file for the policy manager is provided.
Currently, the rtmgr will only partially parse the policy directives. The
more complex parsing is done the the policy process itself. This will be fixed
in the future.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@4635 26435e26-4b25-483a-942d-477236992e84